refactor: strategy folder is now strategies
