+++ lua/codecompanion/actions.lua
@@ -505,14 +505,14 @@ M.static.actions = {
     strategy = "saved_chats",
     description = "Load your previously saved chats",
     condition = function()
-      local saved_chats = require("codecompanion.strategy.saved_chats")
+      local saved_chats = require("codecompanion.strategies.saved_chats")
       return saved_chats:has_chats()
     end,
     picker = {
       prompt = "Load chats",
       items = function()
         local client = require("codecompanion").get_client()
-        local saved_chats = require("codecompanion.strategy.saved_chats")
+        local saved_chats = require("codecompanion.strategies.saved_chats")
         local items = saved_chats:list({ sort = true })
 
         local chats = {}
